# Added benchmarks on cacheable child serializers