Relates #1918

Benchmarks for improved serialization plugin

Test | Old, ops/ms | ±, ops/ms |   | New, ops/ms | ±, ops/ms |   | Boost
-- | -- | -- | -- | -- | -- | -- | --
ContextualOverheadBenchmark.decode | 3077.462 | 23.844 |   | 6394.905 | 38.471 |   | 107.80%
ContextualOverheadBenchmark.encode | 4428.256 | 8.527 |   | 15722.271 | 77.355 |   | 255.04%
PolymorphismOverheadBenchmark.base | 10871.82 | 34.643 |   | 11053.065 | 101.819 |   | 1.67%
PolymorphismOverheadBenchmark.poly | 3523.776 | 17.856 |   | 3692.778 | 13.533 |   | 4.80%
PolymorphismOverheadBenchmark.polyChildDecode | 1012.281 | 2.162 |   | 3214.568 | 8.708 |   | 217.56%
PolymorphismOverheadBenchmark.polyChildEncode | 1356.494 | 5.032 |   | 7529.098 | 38.973 |   | 455.04%
UseSerializerOverheadBenchmark.decodeForClass | 7485.336 | 19.212 |   | 8181.806 | 48.467 |   | 9.30%
UseSerializerOverheadBenchmark.decodeForObject | 7586.439 | 21.435 |   | 7877.765 | 13.945 |   | 3.84%
UseSerializerOverheadBenchmark.encodeForClass | 16485.565 | 95.665 |   | 18103.848 | 96.215 |   | 9.82%
UseSerializerOverheadBenchmark.encodeForObject | 12680.371 | 47.56 |   | 13101.02 | 77.441 |   | 3.32%